requirements.txt
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
django-fsm==2.8.0
django-model-utils==4.2.0
django-modeltranslation==0.18.2
django-pint==0.6.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-constant-field==0.9.1
djangorestframework-serializer-extensions==2.0.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.24.2
drf-writable-nested==0.6.3
drf-yasg==1.21.4
executing==0.8.2
graphviz==0.19.1
gunicorn==20.1.0
hashids==1.3.1
idna==3.3
inflection==0.5.1
ipython==8.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.23.3
openapi-codec==1.3.2
packaging==21.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pint==0.18
platformdirs==2.4.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==1.2.2
traitlets==5.1.1
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
whitenoise==5.3.0
